Use DefaultMethodItem.legacyDefaultValue() for Psi Java
Previously, `PsiMethodItem` would provide its own implementation of
`legacyDefaultValue()` for Java and Kotlin. However, the
`DefaultMethodItem.legacyDefaultValue()` has equivalent behavior for
Java so this change removes the Psi code for Java.

This required a change to a test that used a cast in the default
expression. Previously, it would output the cast to the signature file
and stubs but after this change it just outputs the value. So,
`(double)Float.NEGATIVE_INFINITY` becomes `Double.NEGATIVE_INFINITY`.
The reason that happens is because the cast expression is converted to
its numerical value of `(-1.0/0.0)` and then on output it is mapped to
the field `Double.NEGATIVE_INFINITY`. This does not actually affect
either Android or AndroidX as neither uses a cast in this way.

Bug: 354633349
Test: ./gradlew
      # Tested in Android and AndroidX and does not affect signature
      # files in either or stubs in the former.

+++ metalava/src/test/java/com/android/tools/metalava/stub/StubsAnnotationTest.kt
@@ -463,7 +463,7 @@ class StubsAnnotationTest : AbstractStubsTest() {
                     method public abstract float floating() default 1.0f;
                     method public abstract boolean formatToHexString() default false;
                     method public abstract double from() default java.lang.Double.NEGATIVE_INFINITY;
-                    method public abstract double fromWithCast() default (double)java.lang.Float.NEGATIVE_INFINITY;
+                    method public abstract double fromWithCast() default java.lang.Double.NEGATIVE_INFINITY;
                     method public abstract boolean hasAdjacentMapping() default false;
                     method public abstract int integer() default 1;
                     method public abstract double large_floating() default 1.0;
@@ -503,7 +503,7 @@ class StubsAnnotationTest : AbstractStubsTest() {
                 public float floating() default 1.0f;
                 public boolean formatToHexString() default false;
                 public double from() default java.lang.Double.NEGATIVE_INFINITY;
-                public double fromWithCast() default (double)java.lang.Float.NEGATIVE_INFINITY;
+                public double fromWithCast() default java.lang.Double.NEGATIVE_INFINITY;
                 public boolean hasAdjacentMapping() default false;
                 public int integer() default 1;
                 public double large_floating() default 1.0;
